# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QR Digital Menu API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Restaurant {
  id: number;
  name: string;
  description?: string | null;
  logoUrl?: string | null;
  phone?: string | null;
  address?: string | null;
  primaryColor?: string | null;
  slug: string;
  createdAt: string;
}

export interface CreateRestaurantInput {
  name: string;
  description?: string | null;
  logoUrl?: string | null;
  phone?: string | null;
  address?: string | null;
  primaryColor?: string | null;
  slug: string;
}

export interface Category {
  id: number;
  restaurantId: number;
  name: string;
  imageUrl?: string | null;
  sortOrder: number;
}

export interface CreateCategoryInput {
  name: string;
  imageUrl?: string | null;
  sortOrder: number;
}

export type MenuItemType = (typeof MenuItemType)[keyof typeof MenuItemType];

export const MenuItemType = {
  veg: "veg",
  "non-veg": "non-veg",
  beverage: "beverage",
} as const;

export interface MenuItem {
  id: number;
  restaurantId: number;
  categoryId?: number | null;
  name: string;
  description?: string | null;
  price: number;
  imageUrl?: string | null;
  type: MenuItemType;
  isCombo: boolean;
  comboItemCount?: number | null;
  isSpicy: boolean;
  isAvailable: boolean;
  sortOrder: number;
  categoryName?: string | null;
}

export type CreateMenuItemInputType =
  (typeof CreateMenuItemInputType)[keyof typeof CreateMenuItemInputType];

export const CreateMenuItemInputType = {
  veg: "veg",
  "non-veg": "non-veg",
  beverage: "beverage",
} as const;

export interface CreateMenuItemInput {
  categoryId?: number | null;
  name: string;
  description?: string | null;
  price: number;
  imageUrl?: string | null;
  type: CreateMenuItemInputType;
  isCombo: boolean;
  comboItemCount?: number | null;
  isSpicy: boolean;
  isAvailable: boolean;
  sortOrder: number;
}

export interface Offer {
  id: number;
  restaurantId: number;
  title: string;
  description?: string | null;
  imageUrl?: string | null;
  discountPercent?: number | null;
  validUntil?: string | null;
  isActive: boolean;
}

export interface CreateOfferInput {
  title: string;
  description?: string | null;
  imageUrl?: string | null;
  discountPercent?: number | null;
  validUntil?: string | null;
  isActive: boolean;
}

export interface FullMenu {
  restaurant: Restaurant;
  categories: Category[];
  items: MenuItem[];
  offers: Offer[];
}

export interface QrCodeData {
  url: string;
  restaurantId: number;
  restaurantName: string;
}

export interface RestaurantStats {
  totalCategories: number;
  totalItems: number;
  totalOffers: number;
  vegCount: number;
  nonVegCount: number;
  beverageCount: number;
  comboCount: number;
}

export type ListItemsParams = {
  categoryId?: number;
  type?: ListItemsType;
  search?: string;
};

export type ListItemsType = (typeof ListItemsType)[keyof typeof ListItemsType];

export const ListItemsType = {
  veg: "veg",
  "non-veg": "non-veg",
  beverage: "beverage",
} as const;