# lib/api-zod/src/generated/types/qrCodeData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QR Digital Menu API
 * OpenAPI spec version: 0.1.0
 */

export interface QrCodeData {
  url: string;
  restaurantId: number;
  restaurantName: string;
}